Move static files serve functionality to `Resources`
