+++ lib/application.js
@@ -224,33 +224,6 @@ class Application extends events.EventEmitter {
     }
     return intro;
   }
-
-  serveStatic(url, transport) {
-    const [urlPath, params] = metautil.split(url, '?');
-    const folder = urlPath.endsWith('/');
-    const filePath = urlPath + (folder ? 'index.html' : '');
-    const fileExt = metautil.fileExt(filePath);
-    const data = this.static.get(filePath);
-    if (Buffer.isBuffer(data)) {
-      transport.write(data, 200, fileExt);
-      return;
-    }
-    if (!folder && this.static.get(urlPath + '/index.html')) {
-      const query = params ? '?' + params : '';
-      transport.redirect(urlPath + '/' + query);
-      return;
-    }
-    const absPath = path.join(this.static.path, url);
-    if (absPath.startsWith(this.static.path)) {
-      const readable = fs.createReadStream(absPath);
-      readable.on('error', () => {
-        transport.error(404);
-      });
-      transport.write(readable, 200, fileExt);
-      return;
-    }
-    transport.error(404);
-  }
 }
 
 module.exports = new Application();
