+++ lib/resources.js
@@ -1,7 +1,8 @@
 'use strict';
 
 const path = require('node:path');
-const fsp = require('node:fs').promises;
+const fs = require('node:fs');
+const fsp = fs.promises;
 const metautil = require('metautil');
 const { Cache } = require('./cache.js');
 
@@ -50,6 +51,34 @@ class Resources extends Cache {
     } catch {
       this.delete(filePath);
     }
+  }
+
+  serve(url, transport) {
+    const [urlPath, params] = metautil.split(url, '?');
+    const folder = urlPath.endsWith('/');
+    const filePath = urlPath + (folder ? 'index.html' : '');
+    const fileExt = metautil.fileExt(filePath);
+    const data = this.get(filePath);
+    if (Buffer.isBuffer(data)) {
+      transport.write(data, 200, fileExt);
+      return;
+    }
+    if (!folder && this.get(urlPath + '/index.html')) {
+      const query = params ? '?' + params : '';
+      transport.redirect(urlPath + '/' + query);
+      return;
+    }
+    const absPath = path.join(this.path, url);
+    if (absPath.startsWith(this.path)) {
+      const options = metautil.parseRange(transport.req.headers.range);
+      const readable = fs.createReadStream(absPath, options);
+      readable.on('error', () => {
+        transport.error(404);
+      });
+      transport.write(readable, 200, fileExt, options);
+      return;
+    }
+    transport.error(404);
   }
 }
